deduplicate contract indexer logic
